# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Strata — Unified Asset & Operations Brain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /knowledge-graph:
    get:
      operationId: getKnowledgeGraph
      tags: [knowledge-graph]
      summary: Get the full knowledge graph (nodes + edges)
      parameters:
        - name: entityType
          in: query
          schema:
            $ref: "#/components/schemas/EntityType"
      responses:
        "200":
          description: Knowledge graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGraph"
  /entities/{id}:
    get:
      operationId: getEntity
      tags: [knowledge-graph]
      summary: Get a single entity with its relations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /maintenance/work-orders:
    get:
      operationId: listWorkOrders
      tags: [maintenance]
      summary: List maintenance work orders
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/WorkOrderStatus"
      responses:
        "200":
          description: List of work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
...
  /maintenance/rca/generate:
    post:
      operationId: generateRcaInsights
      tags: [maintenance]
      summary: Run the RCA agent over open/unresolved work orders
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RcaGenerateInput"
      responses:
        "200":
          description: Generated RCA insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RcaInsight"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    EntityType:
      type: string
      enum:
        - equipment
        - procedure
        - regulation
        - personnel
        - location
        - hazard
        - parameter
    Entity:
      type: object
      required: [id, entityType, name, description, documentId]
      properties:
        id: { type: integer }
        entityType:
          $ref: "#/components/schemas/EntityType"
        name: { type: string }
        description: { type: ["string", "null"] }
        documentId: { type: ["integer", "null"] }
    EntityRelation:
      type: object
      required: [id, sourceEntityId, targetEntityId, relationType]
      properties:
        id: { type: integer }
        sourceEntityId: { type: integer }
        targetEntityId: { type: integer }
        relationType: { type: string }
        description: { type: ["string", "null"] }
    EntityDetail:
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          required: [outgoingRelations, incomingRelations]
          properties:
            outgoingRelations:
              type: array
              items:
                $ref: "#/components/schemas/EntityRelation"
            incomingRelations:
              type: array
              items:
                $ref: "#/components/schemas/EntityRelation"
    KnowledgeGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/EntityRelation"
...
    WorkOrderStatus:
      type: string
      enum: [open, in_progress, completed]
    WorkOrderPriority:
      type: string
      enum: [low, medium, high, critical]
    WorkOrder:
      type: object
      required:
        [id, equipmentName, description, status, priority, reportedDate, documentId, createdAt]
      properties:
        id: { type: integer }
        equipmentName: { type: string }
        description: { type: string }
        status:
          $ref: "#/components/schemas/WorkOrderStatus"
        priority:
          $ref: "#/components/schemas/WorkOrderPriority"
        reportedDate: { type: string, format: date }
        resolvedDate: { type: ["string", "null"], format: date }
        documentId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    RcaInsight:
      type: object
      required:
        [id, workOrderId, equipmentName, title, rootCause, contributingFactors, recommendations, createdAt]
      properties:
        id: { type: integer }
        workOrderId: { type: ["integer", "null"] }
        equipmentName: { type: string }
        title: { type: string }
        rootCause: { type: string }
        contributingFactors:
          type: array
          items: { type: string }
        recommendations:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
    RcaGenerateInput:
      type: object
      properties:
        workOrderId: { type: integer }
...